list_test.go core.go entries.go entry.go list.go list_test.go
