--- src/cli/dialog/components/bubble_list/list_test.go
@@ -1,8 +1,0 @@
-package bubblelist_test
-
-import "testing"
-
-func TestBubbleList(t *testing.T) {
-	t.Parallel()
-	t.Run("", func(t *testing.T) {})
-}
