+++ src/cli/dialog/components/list/core.go
@@ -1,0 +1,2 @@
+// Package list provides a generic list implementation to be used inside BubbleTea-based dialog components.
+package list
