+++ src/cli/dialog/components/list/entry.go
@@ -2,8 +2,8 @@ package list
 
 import "fmt"
 
+// Entry is an entry in a List instance.
 type Entry[S fmt.Stringer] struct {
-	// TODO: Checked bool in a RadioListEntry struct
 	Data    S
 	Enabled bool
 	Text    string
