+++ src/cli/dialog/components/list/list_test.go
@@ -2,7 +2,7 @@ package list_test
 
 import "testing"
 
-func TestBubbleList(t *testing.T) {
+func TestList(t *testing.T) {
 	t.Parallel()
 	t.Run("", func(t *testing.T) {
 		t.Parallel()
